refactor: 코드리뷰 1차 수정
