+++ src/main/kotlin/views/Output.kt
@@ -9,7 +9,7 @@ object Output {
         println("지뢰찾기 게임 시작")
         (0 until gameSize.width).forEach { width ->
             (0 until gameSize.height).forEach { height ->
-                val block = gameMap.value.single { it.position == Position(width, height) }
+                val block = gameMap.getBlockByPosition(Position(width,height))
                 print("${block.marker} ")
             }
             println()
